Model/requirements.txt
Flask
gunicorn
tensorflow
Pillow